fix: file router, divide to components
