+++ src/app/courses/[courseId]/page.tsx
@@ -1,0 +1,9 @@
+import DetailCourse from '@/pages/CourseDetailPage';
+
+export default function CourseDetailPage() {
+  return (
+    <div>
+      <DetailCourse />
+    </div>
+  );
+}
